redo this thing

+++ Aika.py
@@ -293,45 +293,58 @@ async def on_message(message):
             """
             if messagecontent[0].lower() == '$user' \
             or messagecontent[0].lower() == '$stats': # Akatsuki userinfo command
-                userID = messagecontent[1]
+                username = messagecontent[1]
                 try:
                     relax = messagecontent[2]
                 except:
                     relax = ''
 
-                processingMessage = await client.send_message(message.channel, 'Processing request...')
+                if relax != '' and relax != '-rx':
+                    await client.send_message(message.channel, "Please use underscores in your username rather than spaces.")
+                else:
+                    try:
+                        processingMessage = await client.send_message(message.channel, 'Processing request...')
 
-                resp = requests.get('https://akatsuki.pw/api/v1/users/{rx}full?id={userID}'.format(rx="rx" if relax == '-rx' else '', userID=userID), timeout=3).text
-                
-                userInfo = json.loads(resp)
-                
-                if userInfo["favourite_mode"] == 0: # osu!
-                    mode = 'std'
-                    modeNice = 'osu!'
-                elif userInfo["favourite_mode"] == 1: # osu!taiko
-                    mode = 'taiko'
-                    modeNice = 'osu!taiko'
-                elif userInfo["favourite_mode"] == 2: # osu!catch
-                    mode = 'ctb'
-                    modeNice = 'osu!catch'
-                elif userInfo["favourite_mode"] == 3: # osu!mania
-                    mode = 'mania'
-                    modeNice = 'osu!mania'
+                        gamer = requests.get('https://akatsuki.pw/api/v1/get_user?u={}'.format(username)).text
 
-                embed = discord.Embed(title="{flag} {username} | {gm} {rx}".format(flag=":flag_{}:".format(userInfo["country"].lower()), username=userInfo["username"], rx='(Relax)' if relax == '-rx' else '(Vanilla)', gm=modeNice), description='** **', color=0x00ff00)
-                embed.set_thumbnail(url='https://i.namir.in/Mbp.png')
+                        gamerInfo = json.loads(gamer)
 
-                embed.add_field(name="Global Rank", value="#{:,}".format(userInfo["{}".format(mode)]["global_leaderboard_rank"]), inline=True)
-                embed.add_field(name="Country Rank", value="#{:,}".format(userInfo["{}".format(mode)]["country_leaderboard_rank"]), inline=True)
-                embed.add_field(name="PP", value="{:,}pp".format(userInfo["{}".format(mode)]["pp"]), inline=True)
-                embed.add_field(name="Ranked Score", value="{:,}".format(userInfo["{}".format(mode)]["ranked_score"]), inline=True)
-                embed.add_field(name="Total Score", value="{:,}".format(userInfo["{}".format(mode)]["total_score"]), inline=True)
-                embed.add_field(name="Accuracy", value=userInfo["{}".format(mode)]["accuracy"], inline=True)
-                embed.add_field(name="Playcount", value="{:,}".format(userInfo["{}".format(mode)]["playcount"]), inline=True)
-                embed.add_field(name="Replays Watched", value="{:,}".format(userInfo["{}".format(mode)]["replays_watched"]), inline=True)
+                        userID = int(gamerInfo[0]["user_id"])
 
-                await client.send_message(message.channel, embed=embed)
-                await client.delete_message(processingMessage)
+                        resp = requests.get('https://akatsuki.pw/api/v1/users/{rx}full?id={userID}'.format(rx="rx" if relax == '-rx' else '', userID=userID), timeout=3).text
+                        
+                        userInfo = json.loads(resp)
+
+                        if userInfo["favourite_mode"] == 0: # osu!
+                            mode = 'std'
+                            modeNice = 'osu!'
+                        elif userInfo["favourite_mode"] == 1: # osu!taiko
+                            mode = 'taiko'
+                            modeNice = 'osu!taiko'
+                        elif userInfo["favourite_mode"] == 2: # osu!catch
+                            mode = 'ctb'
+                            modeNice = 'osu!catch'
+                        elif userInfo["favourite_mode"] == 3: # osu!mania
+                            mode = 'mania'
+                            modeNice = 'osu!mania'
+
+                        embed = discord.Embed(title="{flag} {username} | {gm} {rx}".format(flag=":flag_{}:".format(userInfo["country"].lower()), username=userInfo["username"], rx='(Relax)' if relax == '-rx' else '(Vanilla)', gm=modeNice), description='** **', color=0x00ff00)
+                        embed.set_thumbnail(url='https://i.namir.in/Mbp.png')
+
+                        embed.add_field(name="Global Rank", value="#{:,}".format(userInfo["{}".format(mode)]["global_leaderboard_rank"]), inline=True)
+                        embed.add_field(name="Country Rank", value="#{:,}".format(userInfo["{}".format(mode)]["country_leaderboard_rank"]), inline=True)
+                        embed.add_field(name="PP", value="{:,}pp".format(userInfo["{}".format(mode)]["pp"]), inline=True)
+                        embed.add_field(name="Ranked Score", value="{:,}".format(userInfo["{}".format(mode)]["ranked_score"]), inline=True)
+                        embed.add_field(name="Total Score", value="{:,}".format(userInfo["{}".format(mode)]["total_score"]), inline=True)
+                        embed.add_field(name="Accuracy", value=userInfo["{}".format(mode)]["accuracy"], inline=True)
+                        embed.add_field(name="Playcount", value="{:,}".format(userInfo["{}".format(mode)]["playcount"]), inline=True)
+                        embed.add_field(name="Replays Watched", value="{:,}".format(userInfo["{}".format(mode)]["replays_watched"]), inline=True)
+
+                        await client.send_message(message.channel, embed=embed)
+                        await client.delete_message(processingMessage)
+                    except:
+                        await client.send_message(message.channel, "Either that user does not exist, or the format of your message was incorrect. Format: | $stats username_spaced_like_this -rx\n\n(-rx is optional and will return -rx stats)")
+
 
             elif messagecontent[0].lower() == '$d' or messagecontent[0].lower() == 'psychedelic' or messagecontent[0].lower() == 'psych': # d
                 if random.randint(1, 2) == 1:
